--- model/card_support.go
@@ -1,6 +1,0 @@
-package model
-
-type CardSupport struct {
-	Card    *Card   `json:"card"`
-	Support *[]Card `json:"support"`
-}
