+++ model/deck_list.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/ygo-skc/skc-suggestion-engine/util"
+	"github.com/ygo-skc/skc-suggestion-engine/validation"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
@@ -48,8 +48,8 @@ type DeckListBreakdown struct {
 
 // validate and handle validation error messages
 func (dl DeckList) Validate() *APIError {
-	if err := util.V.Struct(dl); err != nil {
-		return &APIError{Message: util.HandleValidationErrors(err)}
+	if err := validation.V.Struct(dl); err != nil {
+		return &APIError{Message: validation.HandleValidationErrors(err)}
 	} else {
 		return nil
 	}
